build: update GitHub Actions
Update GitHub Actions to their latest versions and pin them using hashes
to prevent tampering. Tags are added as comments and both tags and
hashes will be updated by Dependabot in future pull requests.

Add `--frozen-lockfile` to `pnpm install`s to make it explicit that this
is what is happening in CI.